Factored out explicit cast checks from implicit ones. Now we check for both. Test cases for implicit casts added. reallocarray somehow does not lead to implicit casts

+++ tests/typecast-bugs/realloc-implicit.c
@@ -5,6 +5,6 @@ void f() {
    unsigned long j;
    i = j = 10;
    void *ptr = malloc(j);
-   ptr = realloc(ptr, (unsigned)i);
+   ptr = realloc(ptr, i);
    free(ptr);
 }
